Create To Do List
This script implements a simple to-do list application with the following functionalities:

Adding a task
Listing all tasks
Marking a task as done
Exiting the application
It uses a JSON file (tasks.json) to store tasks and their completion status. The tasks are stored as a list of dictionaries, where each dictionary represents a task with keys 'task' (the task description) and 'completed' (a boolean indicating whether the task is completed or not).